LayoutTests/imported/w3c/web-platform-tests/trusted-types/csp-block-eval-expected.txt

CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"

PASS eval with plain string throws (both block).
PASS eval with TrustedScript throws (script-src blocks).
